backend/requirements.txt
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0
langchain-openai>=0.0.2

# RAG and Document Processing
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.20

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Resume Analysis - File Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
python-multipart>=0.0.6
